feat: change .js extension to .ts
